better script

+++ make-and-upload-docs.sh
@@ -2,15 +2,17 @@
 
 echo "Deleting the old docs"
 rm docs/*
-ssh horn@helena.uni-koblenz.de 'rm /home/horn/public_html/funnyqt-docs/*'
 
-lein marg
-lein html5-docs
+lein marg && lein html5-docs
 
-scp -r docs/* horn@helena.uni-koblenz.de:/home/horn/public_html/funnyqt-docs/
+if [[ $? -eq 0 ]]; then
+    ssh horn@helena.uni-koblenz.de 'rm /home/horn/public_html/funnyqt-docs/*'
+    scp -r docs/* horn@helena.uni-koblenz.de:/home/horn/public_html/funnyqt-docs/
+    ssh horn@helena.uni-koblenz.de \
+	'cd /home/horn/public_html/funnyqt-docs && chmod -R 644 *'
+    echo Fini.
+else
+    echo Error.
+fi
 
-ssh horn@helena.uni-koblenz.de \
-    'cd /home/horn/public_html/funnyqt-docs && chmod -R 644 *'
-
-echo Fini.
